perf: better introduction
