+++ docs/geometry/code/scanning/scanning_4.cpp
@@ -1,31 +1,27 @@
 #include <algorithm>
 #include <iostream>
+using ll = long long;
+using namespace std;
 
-int n, m, a[1000010], ans[1000010];
-int pre[1000010], lst[1000010];  // 处理 pre
+struct node {
+  ll data;
+  ll num;
+} f[500010];
 
-struct ope {
-  int type, x, y, id;
+ll n, ans, a[500010];
 
-  ope(int type = 0, int x = 0, int y = 0, int id = 0) {
-    this->type = type, this->x = x, this->y = y, this->id = id;
+bool cmp(node a, node b) {
+  if (a.data == b.data) {
+    return a.num < b.num;
   }
+  return a.data < b.data;
+}
 
-  bool operator<(const ope& rhs) const {
-    if (x == rhs.x) return type < rhs.type;
-    return x < rhs.x;
-  }
-};
-
-ope op[2500010];
-int tot;  // 操作总数
-
-int sum[1000010];  // 树状数组
+ll sum[500010];
 
 int lowbit(int x) { return x & (-x); }
 
 void add(int x, int k) {
-  x++;  // 位置 0 也要进行修改，所以树状数组下标均加 1
   while (x <= n) {
     sum[x] = sum[x] + k;
     x = x + lowbit(x);
@@ -33,7 +29,6 @@ void add(int x, int k) {
 }
 
 int getsum(int x) {
-  x++;
   int ret = 0;
   while (x > 0) {
     ret = ret + sum[x];
@@ -42,32 +37,20 @@ int getsum(int x) {
   return ret;
 }
 
-using std::cin;
-using std::cout;
-
 int main() {
-  cin.tie(nullptr)->sync_with_stdio(false);
   cin >> n;
+  for (ll i = 1; i <= n; i++) {
+    cin >> f[i].data;
+    f[i].num = i;
+  }
+  sort(f + 1, f + 1 + n, cmp);
   for (int i = 1; i <= n; i++) {
-    cin >> a[i];
-    pre[i] = lst[a[i]], lst[a[i]] = i;  // 处理 pre
-    op[++tot] = ope{0, i, pre[i], i};   // 加点操作
+    a[f[i].num] = i;
   }
-  cin >> m;
-  for (int i = 1, l, r; i <= m; i++) {
-    cin >> l >> r;
-    op[++tot] = ope{1, r, l - 1, i};  // 将查询差分
-    op[++tot] = ope{2, l - 1, l - 1, i};
+  for (ll i = n; i > 0; i--) {
+    ans += getsum(a[i]);
+    add(a[i], 1);
   }
-  std::sort(op + 1, op + tot + 1);  // 将操作按横坐标排序，且优先执行加点操作
-  for (int i = 1; i <= tot; i++) {
-    if (op[i].type == 0)
-      add(op[i].y, 1);
-    else if (op[i].type == 1)
-      ans[op[i].id] += getsum(op[i].y);
-    else
-      ans[op[i].id] -= getsum(op[i].y);
-  }
-  for (int i = 1; i <= m; i++) cout << ans[i] << '\n';
+  cout << ans;
   return 0;
 }
