+++ docs/geometry/examples/scanning/scanning_2.ans
@@ -1,1 +1,3 @@
-3
+Test case #1
+Total explored area: 180.00
+
