+++ docs/geometry/examples/scanning/scanning_2.in
@@ -1,5 +1,4 @@
-3 1
-0 0 
-0 1
-1 0
-0 0 1 1
+2
+10 10 20 20
+15 15 25 25.5
+0
