+++ docs/geometry/examples/scanning/scanning_3.ans
@@ -1,1 +1,1 @@
-11
+3
