+++ docs/geometry/examples/scanning/scanning_3.in
@@ -1,2 +1,5 @@
-6
-5 4 2 6 3 1
+3 1
+0 0 
+0 1
+1 0
+0 0 1 1
