+++ docs/geometry/examples/scanning/scanning_4.ans
@@ -1,3 +1,1 @@
-2
-2
-4
+11
