+++ docs/geometry/examples/scanning/scanning_4.in
@@ -1,6 +1,2 @@
 6
-1 2 3 4 3 5
-3
-1 2
-3 5
-2 6
+5 4 2 6 3 1
