+++ docs/geometry/examples/scanning/scanning_5.ans
@@ -1,0 +1,3 @@
+2
+2
+4
