+++ docs/geometry/examples/scanning/scanning_5.in
@@ -1,0 +1,6 @@
+6
+1 2 3 4 3 5
+3
+1 2
+3 5
+2 6
