+++ docs/geometry/scanning.md
@@ -2,17 +2,13 @@
 
 扫描线一般运用在图形上面，它和它的字面意思十分相似，就是一条线在整个图上扫来扫去，它一般被用来解决图形面积，周长，以及二维数点等问题。
 
-## Atlantis 问题
-
-### 题意
+## 二维矩形面积并问题
 
 在二维坐标系上，给出多个矩形的左下以及右上坐标，求出所有矩形构成的图形的面积。
 
-### 解法
+### 过程
 
 根据图片可知总面积可以直接暴力即可求出面积，如果数据大了怎么办？这时就需要讲到 **扫描线** 算法。
-
-### 过程
 
 现在假设我们有一根线，从下往上开始扫描：
 
@@ -22,7 +18,8 @@
 
 给每一个矩形的上下边进行标记，下面的边标记为 1，上面的边标记为 -1。每遇到一个水平边时，让这条边（在横轴投影区间）的权值加上这条边的标记。
 
-> 这个操作类似遍历括号序列：开括号加 1，闭括号减 1，「权值」对应当前位置的深度，「权值」是否大于 0，对应当前在不在括号里，也就是这段区间是否记入小矩形的宽度。
+???+ note
+    这个操作类似遍历括号序列：开括号加 1，闭括号减 1，「权值」对应当前位置的深度，「权值」是否大于 0，对应当前在不在括号里，也就是这段区间是否记入小矩形的宽度。
 
 小矩形（不一定只有一个）的宽度就是整个数轴上权值大于 0 的区间总长度。
 
@@ -39,14 +36,18 @@
 
 需要 [离散化](../misc/discrete.md)。
 
-???+ note "模板代码"
+???+ note "[洛谷 P5490【模板】扫描线 & 矩形面积并](https://www.luogu.com.cn/problem/P5490) 参考代码"
     ```cpp
     --8<-- "docs/geometry/code/scanning/scanning_1.cpp"
     ```
 
+???+ note "[「POJ 1151」Atlantis](http://poj.org/problem?id=1151) 参考代码"
+    ```cpp
+    --8<-- "docs/geometry/code/scanning/scanning_2.cpp"
+    ```
+
 ### 练习
 
--   [「POJ1151」Atlantis](http://poj.org/problem?id=1151)
 -   [「POJ1177」Picture](http://poj.org/problem?id=1177)
 -   [「POJ3832」Posters](http://poj.org/problem?id=3832)
 -   [洛谷 P1856 \[IOI1998\] \[USACO5.5\] 矩形周长 Picture](https://www.luogu.com.cn/problem/P1856)
